+++ CHANGES.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- Added tags-vocabulary endpoint to API for index-specific part-of-speech tags.
+- Added tests for metadata and parent API calls in backend.
+- Added metadata query language.  Includes support for dates and nested objects. ([#305](https://github.com/lum-ai/odinson/pull/305))
 ### Changed
+- Parent document filenames are stored by default.
 
 ## [0.4.0] - 2021-06-07
 ### Added